backend/requirements.txt
django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
